+++ packages/sdk-legal/src/custom/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2023 AccelByte Inc. All Rights Reserved
- * This is licensed software from AccelByte Inc, for limitations
- * and restrictions contact your company contract manager.
- */
-export * from './clients/LegalHelper.js'
-export * from './models/Legal.js'