+++ packages/sdk-odinconfig/src/custom/models/DiscoveryTemplate.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2022-2024 AccelByte Inc. All Rights Reserved
- * This is licensed software from AccelByte Inc, for limitations
- * and restrictions contact your company contract manager.
- */
-// This should always be in sync with apps/config-service-app/pkg/constant/constant.go.
-export const DISCOVERY_TEMPLATE_NAME = 'template-discovery'